+++ app/Http/Controllers/JenisBarangController.php
@@ -4,12 +4,17 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\JenisBarang;
+use Illuminate\Support\Facades\Auth;
 
 
 class JenisBarangController extends Controller
 {
     public function index() 
     { 
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+        }
+
         $data = array(
             'title' => 'Jenis Barang',
             'data_jenis_barang' => JenisBarang::all(),
@@ -19,6 +24,10 @@ class JenisBarangController extends Controller
 
     public function store(Request $request)
     {
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+        }
+
         JenisBarang::create([
             'nama_jenis' => $request->nama_jenis,
         ]);
@@ -27,6 +36,10 @@ class JenisBarangController extends Controller
 
     public function update(Request $request, $id)
     {
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+        }
+
             $jenis = JenisBarang::find($id);
             $jenis->update([
             'nama_jenis' => $request->nama_jenis,
@@ -36,6 +49,10 @@ class JenisBarangController extends Controller
 
     public function destroy($id)
     {   
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+        }
+
         $data = JenisBarang::find($id);
         $data->delete();
         return redirect('/jenisbarang')->with('success', 'Data Berhasil');
