+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 class UserController extends Controller
 
 
@@ -11,6 +12,10 @@ class UserController extends Controller
     public function index() {
         // $users = User::all();
         // return response()->json($users);
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+        }
+
         $data = array(
             'title' => 'User',
             'data_user' => User::all(),
@@ -22,6 +27,10 @@ class UserController extends Controller
     public function store(Request $request) {
         // // $user = User::create($request->all());
         // return response()->json($user);
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+        }
+
         User::create([
             'name' => $request->name,
             'email' => $request->email,
@@ -36,6 +45,11 @@ class UserController extends Controller
         // $user = User::findOrFail($id);
         // $user->update($request->all());
         // return response()->json($user);
+
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+        }
+
         $user = User::find($id);
         $user->update([
             'name' => $request->name,
@@ -50,6 +64,10 @@ class UserController extends Controller
     public function destroy($id) {
         // User::destroy($id);
         // return response()->json(['message' => 'User deleted successfully']);
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+        }
+
         $data = User::find($id);
         $data->delete();
         return redirect('/user')->with('success', 'Data Berhasil');
